refactor: cli arguments and src/utils/compiler.py
